+++ examples/structs.rs
@@ -4,7 +4,7 @@
 //  Created:
 //    26 Dec 2024, 12:12:07
 //  Last edited:
-//    04 Feb 2025, 16:36:47
+//    05 Feb 2025, 15:26:49
 //  Auto updated?
 //    Yes
 //