+++ src/partial_eq.rs
@@ -4,7 +4,7 @@
 //  Created:
 //    09 Jan 2025, 01:27:30
 //  Last edited:
-//    09 Jan 2025, 19:26:14
+//    05 Feb 2025, 15:43:11
 //  Auto updated?
 //    Yes
 //
@@ -20,7 +20,7 @@ use syn::punctuated::Punctuated;
 use syn::spanned::Spanned as _;
 use syn::{Data, DeriveInput, Field, Fields, Ident, LitInt, Path, PathArguments, PathSegment, Token, parse_macro_input};
 
-use crate::extract::extract_generics;
+use crate::common::{extract_generics, filter_skipped_variants_and_fields};
 
 
 /***** HELPER FUNCTIONS *****/
@@ -174,7 +174,12 @@ fn build_eq_impl(input: &DeriveInput) -> TokenStream2 {
 /// # Returns
 /// A [`TokenSream2`] encoding the impl.
 pub fn partial_eq(input: TokenStream) -> TokenStream {
-    let input = parse_macro_input!(input as DeriveInput);
+    let mut input = parse_macro_input!(input as DeriveInput);
+
+    // Filter the input data
+    if let Err(err) = filter_skipped_variants_and_fields("partial_eq", &mut input.data) {
+        return err.into_compile_error().into();
+    }
 
     // Extract the generics & fmts for the general impl
     let generics = extract_generics(&input, &Path {